[UPDATE] huffman uncompression for txt files

+++ include/mementar/lz/BitFileGenerator.h
@@ -9,11 +9,11 @@ class BitFileGenerator
 public:
   BitFileGenerator(size_t size_1 = 0, size_t size_2 = 0, size_t size_3 = 0, size_t size_4 = 0);
 
-  void writeType1(uint16_t value);
-  void writeType2(uint16_t value);
-  void writeType3(uint16_t value);
-  void writeType4(uint16_t value);
-  void writeN(size_t size, uint16_t value);
+  void writeType1(uint32_t value);
+  void writeType2(uint32_t value);
+  void writeType3(uint32_t value);
+  void writeType4(uint32_t value);
+  void writeN(size_t size, uint32_t value);
   void writeChar(char value);
 
   void writeBitTrue();
