+++ src/lz/BitFileGetter.cpp
@@ -13,13 +13,13 @@ BitFileGetter::BitFileGetter(size_t size_1, size_t size_2, size_t size_3, size_t
   minor_index_ = 0;
 }
 
-uint16_t BitFileGetter::getType1()
+uint32_t BitFileGetter::getType1()
 {
-  uint16_t res = 0;
+  uint32_t res = 0;
   int16_t to_get = type_1_size_;
   for(size_t i = 0; i < type_1_size_;)
   {
-    uint16_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
+    uint32_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
     res = res | ((working_data & ~((0xffff) << to_get)) << (type_1_size_ - to_get));
 
     uint8_t getted = (minor_index_ + to_get >= 8) ? 8 - minor_index_ : to_get;
@@ -45,13 +45,13 @@ uint16_t BitFileGetter::getType1()
   return res;
 }
 
-uint16_t BitFileGetter::getType2()
+uint32_t BitFileGetter::getType2()
 {
-  uint16_t res = 0;
+  uint32_t res = 0;
   int16_t to_get = type_2_size_;
   for(size_t i = 0; i < type_2_size_;)
   {
-    uint16_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
+    uint32_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
     res = res | ((working_data & ~((0xffff) << to_get)) << (type_2_size_ - to_get));
 
     uint8_t getted = (minor_index_ + to_get >= 8) ? 8 - minor_index_ : to_get;
@@ -77,13 +77,13 @@ uint16_t BitFileGetter::getType2()
   return res;
 }
 
-uint16_t BitFileGetter::getType3()
+uint32_t BitFileGetter::getType3()
 {
-  uint16_t res = 0;
+  uint32_t res = 0;
   int16_t to_get = type_3_size_;
   for(size_t i = 0; i < type_3_size_;)
   {
-    uint16_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
+    uint32_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
     res = res | ((working_data & ~((0xffff) << to_get)) << (type_3_size_ - to_get));
 
     uint8_t getted = (minor_index_ + to_get >= 8) ? 8 - minor_index_ : to_get;
@@ -109,13 +109,13 @@ uint16_t BitFileGetter::getType3()
   return res;
 }
 
-uint16_t BitFileGetter::getType4()
+uint32_t BitFileGetter::getType4()
 {
-  uint16_t res = 0;
+  uint32_t res = 0;
   int16_t to_get = type_4_size_;
   for(size_t i = 0; i < type_4_size_;)
   {
-    uint16_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
+    uint32_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
     res = res | ((working_data & ~((0xffff) << to_get)) << (type_4_size_ - to_get));
 
     uint8_t getted = (minor_index_ + to_get >= 8) ? 8 - minor_index_ : to_get;
@@ -143,11 +143,11 @@ uint16_t BitFileGetter::getType4()
 
 char BitFileGetter::getChar()
 {
-  uint16_t res = 0;
+  uint32_t res = 0;
   int16_t to_get = 7;
   for(size_t i = 0; i < 7;)
   {
-    uint16_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
+    uint32_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
     res = res | ((working_data & ~((0xffff) << to_get)) << (7 - to_get));
 
     uint8_t getted = (minor_index_ + to_get >= 8) ? 8 - minor_index_ : to_get;
@@ -177,7 +177,7 @@ bool BitFileGetter::getBit()
 {
   bool res = false;
 
-  uint16_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
+  uint32_t working_data = (data_[major_index_] & 0x00ff) >> minor_index_;
   res = working_data & 0x01;
 
   minor_index_ ++;
